requirements.txt
# ML & Data
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0
joblib>=1.3.0

# Audio feature extraction
praat-parselmouth>=0.4.3
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data fetching
ucimlrepo>=0.0.3

# Plotting (model evaluation)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.0.0
httpx>=0.24.0
